+++ db/lib/Property.go
@@ -24,7 +24,7 @@ type Property struct {
 	DownPayment       float64
 	RentableArea      int64
 	RentableAreaUnits int64
-	LotSize           int64
+	LotSize           float64
 	LotSizeUnits      int64
 	CapRate           float64
 	AvgCap            float64
